refactoring of limits handler
