+++ network-store-server/src/main/java/com/powsybl/network/store/server/QueryCatalog.java
@@ -9,7 +9,6 @@ package com.powsybl.network.store.server;
 import com.powsybl.network.store.model.Resource;
 
 import java.util.Collection;
-import java.util.Collections;
 import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
@@ -42,34 +41,20 @@ public final class QueryCatalog {
     public static final String EQUIPMENT_ID_COLUMN = "equipmentId";
     public static final String AREA_ID_COLUMN = "areaId";
     static final String REGULATING_EQUIPMENT_ID = "regulatingEquipmentId";
-    public static final String INDEX_COLUMN = "index";
     public static final String TAPCHANGER_TYPE_COLUMN = "tapChangerType";
     public static final String TAPCHANGER_STEPS_COLUMN = "tapchangersteps";
     public static final String TAP_CHANGER_TYPE = "tapchangertype";
-    public static final String ALPHA_COLUMN = "alpha";
     public static final String TAP_CHANGER_STEP_TABLE = "tapchangersteps";
     public static final String AREA_BOUNDARY_TABLE = "areaboundary";
     public static final String REACTIVE_CAPABILITY_CURVE_POINT_TABLE = "reactiveCapabilityCurvePoint";
     static final String REGULATING_POINT_TABLE = "regulatingPoint";
     static final String REGULATION_MODE = "regulationMode";
-    public static final String SIDE_COLUMN = "side";
     private static final String TYPE_COLUMN = "type";
     static final String REGULATING = "regulating";
-    public static final String SELECTED_OPERATIONAL_LIMITS_GROUP_ID1 = "selectedoperationallimitsgroupid1";
-    public static final String SELECTED_OPERATIONAL_LIMITS_GROUP_ID2 = "selectedoperationallimitsgroupid2";
     private static final Predicate<String> CLONE_PREDICATE = column -> !column.equals(UUID_COLUMN) && !column.equals(VARIANT_ID_COLUMN)
             && !column.equals(NAME_COLUMN) && !column.equals(FULL_VARIANT_NUM_COLUMN);
     private static final String TOMBSTONED_IDENTIFIABLE_TABLE = "tombstonedidentifiable";
     private static final String TOMBSTONED_EXTERNAL_ATTRIBUTES_TABLE = "tombstonedexternalattributes";
-    static final String OPERATIONAL_LIMITS_GROUP_TABLE = "operationallimitsgroup";
-    static final String GROUP_ID_COLUMN = "operationallimitgroupid";
-    static final String CURRENT_LIMITS_PERMANENT_LIMIT_COLUMN = "current_limits_permanent_limit";
-    static final String CURRENT_LIMITS_TEMPORARY_LIMITS_COLUMN = "current_limits_temporary_limits";
-    static final String APPARENT_POWER_LIMITS_PERMANENT_LIMIT_COLUMN = "apparent_power_limits_permanent_limit";
-    static final String APPARENT_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN = "apparent_power_limits_temporary_limits";
-    static final String ACTIVE_POWER_LIMITS_PERMANENT_LIMIT_COLUMN = "active_power_limits_permanent_limit";
-    static final String ACTIVE_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN = "active_power_limits_temporary_limits";
-    static final String PROPERTIES_COLUMN = "properties";
 
     private QueryCatalog() {
     }
@@ -81,15 +66,6 @@ public final class QueryCatalog {
                 " where " + NETWORK_UUID_COLUMN + " = ?" +
                 " and " + VARIANT_NUM_COLUMN + " = ?" +
                 " and " + ID_COLUMN + " = ?";
-    }
-
-    public static String buildGetSelectedOperationalLimitsGroupsQuery(String tableName) {
-        return "select " + ID_COLUMN + ", " +
-            SELECTED_OPERATIONAL_LIMITS_GROUP_ID1 + ", " +
-            SELECTED_OPERATIONAL_LIMITS_GROUP_ID2 +
-            " from " + tableName +
-            " where " + NETWORK_UUID_COLUMN + " = ?" +
-            " and " + VARIANT_NUM_COLUMN + " = ?";
     }
 
     public static String buildGetNetworkQuery(Collection<String> columns) {
@@ -372,35 +348,6 @@ public final class QueryCatalog {
                 VARIANT_NUM_COLUMN + " = ? ";
     }
 
-    // Operational limits
-    public static String buildCloneOperationalLimitsGroupQuery() {
-        return "insert into " + OPERATIONAL_LIMITS_GROUP_TABLE + "(" + EQUIPMENT_ID_COLUMN + ", " + EQUIPMENT_TYPE_COLUMN + ", " +
-                NETWORK_UUID_COLUMN + ", " + VARIANT_NUM_COLUMN + ", " + GROUP_ID_COLUMN + ", " + SIDE_COLUMN + ", " +
-                CURRENT_LIMITS_PERMANENT_LIMIT_COLUMN + ", " + CURRENT_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                APPARENT_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " + APPARENT_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                ACTIVE_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " + ACTIVE_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                PROPERTIES_COLUMN + ") " +
-                "select " + EQUIPMENT_ID_COLUMN + ", " + EQUIPMENT_TYPE_COLUMN + ", ?, ?, " +
-                GROUP_ID_COLUMN + ", " + SIDE_COLUMN + ", " +
-                CURRENT_LIMITS_PERMANENT_LIMIT_COLUMN + ", " + CURRENT_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                APPARENT_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " + APPARENT_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                ACTIVE_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " + ACTIVE_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                PROPERTIES_COLUMN +
-                " from " + OPERATIONAL_LIMITS_GROUP_TABLE + " where " + NETWORK_UUID_COLUMN +
-                " = ? and " + VARIANT_NUM_COLUMN + " = ?";
-    }
-
-    public static String buildDeleteOperationalLimitsGroupQuery() {
-        return "delete from " + OPERATIONAL_LIMITS_GROUP_TABLE + " where " +
-                NETWORK_UUID_COLUMN + " = ?";
-    }
-
-    public static String buildDeleteOperationalLimitsGroupVariantQuery() {
-        return "delete from " + OPERATIONAL_LIMITS_GROUP_TABLE + " where " +
-                NETWORK_UUID_COLUMN + " = ? and " +
-                VARIANT_NUM_COLUMN + " = ?";
-    }
-
     // Tombstoned external attributes
     public static String buildInsertTombstonedExternalAttributesQuery() {
         return "insert into " + TOMBSTONED_EXTERNAL_ATTRIBUTES_TABLE + " (" + NETWORK_UUID_COLUMN + ", " + VARIANT_NUM_COLUMN + ", " + EQUIPMENT_ID_COLUMN + ", " + TYPE_COLUMN + ") " +
@@ -439,31 +386,6 @@ public final class QueryCatalog {
                 "where " +
                 NETWORK_UUID_COLUMN + " = ?" + " and " +
                 VARIANT_NUM_COLUMN + " = ? ";
-    }
-
-    public static String buildDeleteOperationalLimitsGroupVariantEquipmentINQuery(int numberOfValues) {
-        if (numberOfValues < 1) {
-            throw new IllegalArgumentException(MINIMAL_VALUE_REQUIREMENT_ERROR);
-        }
-        return "delete from " + OPERATIONAL_LIMITS_GROUP_TABLE + " where " +
-            NETWORK_UUID_COLUMN + " = ? and " +
-            VARIANT_NUM_COLUMN + " = ? and " +
-            EQUIPMENT_ID_COLUMN + " in (" + generateInPlaceholders(numberOfValues) + ")";
-    }
-
-    public static String buildDeleteOperationalLimitsGroupByGroupIdAndSideAndIdentifiableIdINQuery(int numberOfValues) {
-        if (numberOfValues < 1) {
-            throw new IllegalArgumentException(MINIMAL_VALUE_REQUIREMENT_ERROR);
-        }
-
-        return "delete from " + OPERATIONAL_LIMITS_GROUP_TABLE + " t " +
-                "where t." + NETWORK_UUID_COLUMN + " = ? " +
-                " and t." + VARIANT_NUM_COLUMN + " = ? " +
-                " and exists (select 1 from (values " +
-                String.join(", ", Collections.nCopies(numberOfValues, "(?, ?, ?)")) +
-                ") v(" + EQUIPMENT_ID_COLUMN + ", " + GROUP_ID_COLUMN + ", " + SIDE_COLUMN + ") " +
-                "where (t." + EQUIPMENT_ID_COLUMN + ", t." + GROUP_ID_COLUMN + ", t." + SIDE_COLUMN + ") = " +
-                "      (v." + EQUIPMENT_ID_COLUMN + ", v." + GROUP_ID_COLUMN + ", v." + SIDE_COLUMN + "))";
     }
 
     // Reactive Capability Curve Point
@@ -792,90 +714,5 @@ public final class QueryCatalog {
                 " from " + table + " where " +
                 NETWORK_UUID_COLUMN + " = ? and " +
                 VARIANT_NUM_COLUMN + " = ?";
-    }
-
-    public static String buildInsertOperationalLimitsGroupQuery() {
-        return "insert into " + OPERATIONAL_LIMITS_GROUP_TABLE + " (" +
-            NETWORK_UUID_COLUMN + ", " +
-            VARIANT_NUM_COLUMN + ", " +
-            EQUIPMENT_TYPE_COLUMN + ", " +
-            EQUIPMENT_ID_COLUMN + ", " +
-            GROUP_ID_COLUMN + ", " +
-            SIDE_COLUMN + ", " +
-            CURRENT_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-            CURRENT_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-            APPARENT_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-            APPARENT_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-            ACTIVE_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-            ACTIVE_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-            PROPERTIES_COLUMN + ")" +
-            " values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
-    }
-
-    public static String buildOperationalLimitsGroupQuery(String columnNameForWhereClause) {
-        return "select " + EQUIPMENT_ID_COLUMN + ", " +
-                EQUIPMENT_TYPE_COLUMN + ", " +
-                NETWORK_UUID_COLUMN + ", " +
-                VARIANT_NUM_COLUMN + ", " +
-                SIDE_COLUMN + "," +
-                GROUP_ID_COLUMN + "," +
-                CURRENT_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-                CURRENT_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                APPARENT_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-                APPARENT_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                ACTIVE_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-                ACTIVE_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                PROPERTIES_COLUMN +
-                " from " + OPERATIONAL_LIMITS_GROUP_TABLE + " where " +
-                NETWORK_UUID_COLUMN + " = ? and " +
-                VARIANT_NUM_COLUMN + " = ? and " +
-                columnNameForWhereClause + " = ?";
-    }
-
-    public static String buildOperationalLimitsGroupWithInClauseQuery(String columnNameForInClause, int numberOfValues) {
-        if (numberOfValues < 1) {
-            throw new IllegalArgumentException(MINIMAL_VALUE_REQUIREMENT_ERROR);
-        }
-        return "select " + EQUIPMENT_ID_COLUMN + ", " +
-            EQUIPMENT_TYPE_COLUMN + ", " +
-            NETWORK_UUID_COLUMN + ", " +
-            VARIANT_NUM_COLUMN + ", " +
-            SIDE_COLUMN + "," +
-            GROUP_ID_COLUMN + "," +
-            CURRENT_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-            CURRENT_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-            APPARENT_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-            APPARENT_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-            ACTIVE_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-            ACTIVE_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-            PROPERTIES_COLUMN +
-            " from " + OPERATIONAL_LIMITS_GROUP_TABLE + " where " +
-            NETWORK_UUID_COLUMN + " = ? and " +
-            VARIANT_NUM_COLUMN + " = ? and " +
-            columnNameForInClause + " in (" + generateInPlaceholders(numberOfValues) + ")";
-    }
-
-    public static String buildSelectedOperationalLimitsGroupINQuery(int numberOfValues) {
-        if (numberOfValues < 1) {
-            throw new IllegalArgumentException(MINIMAL_VALUE_REQUIREMENT_ERROR);
-        }
-
-        return "select " + EQUIPMENT_ID_COLUMN + ", " +
-                EQUIPMENT_TYPE_COLUMN + ", " +
-                NETWORK_UUID_COLUMN + ", " +
-                VARIANT_NUM_COLUMN + ", " +
-                SIDE_COLUMN + "," +
-                GROUP_ID_COLUMN + "," +
-                CURRENT_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-                CURRENT_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                APPARENT_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-                APPARENT_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                ACTIVE_POWER_LIMITS_PERMANENT_LIMIT_COLUMN + ", " +
-                ACTIVE_POWER_LIMITS_TEMPORARY_LIMITS_COLUMN + ", " +
-                PROPERTIES_COLUMN +
-                " from " + OPERATIONAL_LIMITS_GROUP_TABLE +
-                " where " + NETWORK_UUID_COLUMN + " = ? and " + VARIANT_NUM_COLUMN + " = ? " +
-                " and (" + EQUIPMENT_ID_COLUMN + ", " + GROUP_ID_COLUMN + ", " + SIDE_COLUMN + ") " +
-                " in (values " + String.join(",", Collections.nCopies(numberOfValues, "(?, ?, ?)")) + ")";
     }
 }
